move treasury things out
