+++ e2e-tests/src/main.rs
@@ -11,10 +11,12 @@ use substrate_api_client::{compose_call, compose_extrinsic, AccountId, XtStatus}
 
 use config::Config;
 
+use crate::treasury::{test_channeling_fee, test_treasury_access};
 use crate::utils::*;
 use crate::waiting::{wait_for_finalized_block, wait_for_session};
 
 mod config;
+mod treasury;
 mod utils;
 mod waiting;
 
@@ -114,74 +116,6 @@ fn test_token_transfer(config: Config) -> anyhow::Result<()> {
         balance_after,
         transfer_value
     );
-
-    Ok(())
-}
-
-// todo: check channeling tips
-fn test_channeling_fee(config: Config) -> anyhow::Result<()> {
-    let (connection, _, to) = setup_for_transfer(config);
-    let treasury = get_treasury_account(&connection);
-
-    let treasury_balance_before = get_free_balance(&treasury, &connection);
-    let issuance_before = get_total_issuance(&connection);
-    info!(
-        "[+] Treasury balance before tx: {}. Total issuance: {}.",
-        treasury_balance_before, issuance_before
-    );
-
-    let tx = transfer(&to, 1000u128, &connection);
-
-    let treasury_balance_after = get_free_balance(&treasury, &connection);
-    let issuance_after = get_total_issuance(&connection);
-    info!(
-        "[+] Treasury balance after tx: {}. Total issuance: {}.",
-        treasury_balance_after, issuance_after
-    );
-
-    assert!(
-        issuance_after <= issuance_before,
-        "Unexpectedly {} was minted",
-        issuance_after - issuance_before
-    );
-    assert!(
-        issuance_before <= issuance_after,
-        "Unexpectedly {} was burned",
-        issuance_before - issuance_after
-    );
-
-    let fee_info = get_tx_fee_info(&connection, &tx);
-    let fee = fee_info.fee_without_weight + fee_info.adjusted_weight;
-
-    assert_eq!(
-        treasury_balance_before + fee,
-        treasury_balance_after,
-        "Incorrect amount was channeled to the treasury: before = {}, after = {}, fee = {}",
-        treasury_balance_before,
-        treasury_balance_after,
-        fee
-    );
-
-    Ok(())
-}
-
-fn test_treasury_access(config: Config) -> anyhow::Result<()> {
-    let Config { node, seeds, .. } = config.clone();
-
-    let proposer = accounts(seeds)[0].to_owned();
-    let beneficiary = AccountId::from(proposer.public());
-    let connection = create_connection(node).set_signer(proposer);
-
-    propose_treasury_spend(0u128, &beneficiary, &connection);
-    propose_treasury_spend(0u128, &beneficiary, &connection);
-    let proposals_counter = get_proposals_counter(&connection);
-    assert!(proposals_counter >= 2, "Proposal was not created");
-
-    let sudo = get_sudo(config);
-    let connection = connection.set_signer(sudo);
-
-    treasury_approve(proposals_counter - 2, &connection)?;
-    treasury_reject(proposals_counter - 1, &connection)?;
 
     Ok(())
 }
